# Simplifies ThrottledCall to reduce GC overhead