--- a/zipkin-server/src/main/java/zipkin2/server/internal/throttle/ThrottledCall.java
+++ b/zipkin-server/src/main/java/zipkin2/server/internal/throttle/ThrottledCall.java
@@ -38,11 +38,21 @@ import zipkin2.Callback;
  * @see ThrottledStorageComponent
  */
 final class ThrottledCall extends Call.Base<Void> {
+  static final Callback<Void> NOOP_CALLBACK = new Callback<Void>() {
+    @Override public void onSuccess(Void value) {
+    }
+
+    @Override public void onError(Throwable t) {
+    }
+  };
+
   final Call<Void> delegate;
   final Executor executor;
   final Limiter<Void> limiter;
   final LimiterMetrics limiterMetrics;
   final Predicate<Throwable> isOverCapacity;
+  final CountDownLatch latch = new CountDownLatch(1);
+  Throwable throwable; // thread visibility guaranteed by the countdown latch
 
   ThrottledCall(Call<Void> delegate, Executor executor, Limiter<Void> limiter,
     LimiterMetrics limiterMetrics, Predicate<Throwable> isOverCapacity) {
@@ -59,38 +69,38 @@ final class ThrottledCall extends Call.Base<Void> {
    * anyway.
    */
   @Override protected Void doExecute() throws IOException {
-    AwaitableCallback awaitableCallback = new AwaitableCallback();
-    doEnqueue(awaitableCallback);
-    if (!await(awaitableCallback.countDown)) throw new InterruptedIOException();
+    // Enqueue the call invocation on the executor and block until it completes.
+    doEnqueue(NOOP_CALLBACK);
+    if (!await(latch)) throw new InterruptedIOException();
 
-    Throwable t = awaitableCallback.throwable;
-    if (t != null) {
-      if (t instanceof Error) throw (Error) t;
-      if (t instanceof IOException) throw (IOException) t;
-      if (t instanceof RuntimeException) throw (RuntimeException) t;
-      throw new RuntimeException(t);
-    }
-    return null; // Void
+    // Check if the run resulted in an exception
+    Throwable t = this.throwable;
+    if (t == null) return null; // success
+
+    // Coerce the throwable to the signature of Call.execute()
+    if (t instanceof Error) throw (Error) t;
+    if (t instanceof IOException) throw (IOException) t;
+    if (t instanceof RuntimeException) throw (RuntimeException) t;
+    throw new RuntimeException(t);
   }
 
+  // When handling enqueue, we don't block the calling thread. Any exception goes to the callback.
   @Override protected void doEnqueue(Callback<Void> callback) {
     Listener limiterListener = limiter.acquire(null)
       .orElseThrow(RejectedExecutionException::new); // TODO: make an exception message
-    limiterMetrics.requests.increment();
-    limiterListener = limiterMetrics.wrap(limiterListener);
 
-    LimiterReleasingCallback releasingCallback =
-      new LimiterReleasingCallback(callback, isOverCapacity, limiterListener);
+    limiterMetrics.requests.increment();
+    EnqueueAndAwait enqueueAndAwait = new EnqueueAndAwait(callback, limiterListener);
 
     try {
-      executor.execute(new EnqueueAndAwait(this, releasingCallback));
+      executor.execute(enqueueAndAwait);
     } catch (RuntimeException | Error t) { // possibly rejected, but from the executor, not storage!
       propagateIfFatal(t);
       callback.onError(t);
       // Ignoring in all cases here because storage itself isn't saying we need to throttle. Though
       // we may still be write bound, but a drop in concurrency won't necessarily help.
       limiterListener.onIgnore();
-      throw t;
+      throw t; // allows blocking calls to see the exception
     }
   }
 
@@ -102,57 +112,72 @@ final class ThrottledCall extends Call.Base<Void> {
     return "Throttled(" + delegate + ")";
   }
 
-  static final class AwaitableCallback implements Callback<Void> {
-    final CountDownLatch countDown = new CountDownLatch(1);
-    Throwable throwable; // thread visibility guaranteed by the countdown latch
+  /** When run, this enqueues a call with a given callback, and awaits its completion. */
+  final class EnqueueAndAwait implements Runnable, Callback<Void> {
+    final Callback<Void> callback;
+    final Listener limiterListener;
 
-    @Override public void onSuccess(Void ignored) {
-      countDown.countDown();
-    }
-
-    @Override public void onError(Throwable t) {
-      throwable = t;
-      countDown.countDown();
-    }
-  }
-
-  static final class EnqueueAndAwait implements Runnable {
-    final Call<Void> delegate;
-    final Predicate<Throwable> isOverCapacity;
-    final LimiterReleasingCallback callback;
-
-    EnqueueAndAwait(ThrottledCall throttledCall, LimiterReleasingCallback callback) {
-      this.delegate = throttledCall.delegate;
-      this.isOverCapacity = throttledCall.isOverCapacity;
+    EnqueueAndAwait(Callback<Void> callback, Listener limiterListener) {
       this.callback = callback;
+      this.limiterListener = limiterListener;
     }
 
     /**
      * This waits until completion to ensure the number of executing calls doesn't surpass the
      * concurrency limit of the executor.
      *
-     * <h3>This component does not affect the {@link Listener} directly</h3>
+     * <h3>The {@link Listener} isn't affected during run</h3>
      * There could be an error enqueuing the call or an interruption during shutdown of the
      * executor. We do not affect the {@link Listener} here because it would be redundant to
-     * handling already done in {@link LimiterReleasingCallback}. For example, if shutting down, the
-     * storage layer would also invoke {@link LimiterReleasingCallback#onError(Throwable)}.
+     * handling already done in callbacks. For example, if shutting down, the storage layer would
+     * also invoke {@link #onError(Throwable)}.
      */
     @Override public void run() {
       if (delegate.isCanceled()) return;
       try {
-        delegate.enqueue(callback);
+        delegate.enqueue(this);
 
         // Need to wait here since the callback call will run asynchronously also.
         // This ensures we don't exceed our throttle/queue limits.
-        await(callback.latch);
+        await(latch);
       } catch (Throwable t) { // edge case: error during enqueue!
         propagateIfFatal(t);
         callback.onError(t);
       }
     }
 
+    @Override public void onSuccess(Void value) {
+      try {
+        // usually we don't add metrics like this,
+        // but for now it is helpful to sanity check acquired vs erred.
+        limiterMetrics.requestsSucceeded.increment();
+        limiterListener.onSuccess(); // NOTE: limiter could block and delay the caller's callback
+        callback.onSuccess(value);
+      } finally {
+        latch.countDown();
+      }
+    }
+
+    @Override public void onError(Throwable t) {
+      try {
+        throwable = t; // catch the throwable in case the invocation is blocking (Call.execute())
+        if (isOverCapacity.test(t)) {
+          limiterMetrics.requestsDropped.increment();
+          limiterListener.onDropped();
+        } else {
+          limiterMetrics.requestsIgnored.increment();
+          limiterListener.onIgnore();
+        }
+
+        // NOTE: the above limiter could block and delay the caller's callback
+        callback.onError(t);
+      } finally {
+        latch.countDown();
+      }
+    }
+
     @Override public String toString() {
-      return "EnqueueAndAwait{call=" + delegate + ", callback=" + callback.delegate + "}";
+      return "EnqueueAndAwait{call=" + delegate + ", callback=" + callback + "}";
     }
   }
 
@@ -177,48 +202,6 @@ final class ThrottledCall extends Call.Base<Void> {
         Thread.currentThread().interrupt();
         return false;
       }
-    }
-  }
-
-  static final class LimiterReleasingCallback implements Callback<Void> {
-    final Callback<Void> delegate;
-    final Predicate<Throwable> isOverCapacity;
-    final Listener limiterListener;
-    final CountDownLatch latch = new CountDownLatch(1);
-
-    LimiterReleasingCallback(Callback<Void> delegate, Predicate<Throwable> isOverCapacity,
-      Listener limiterListener) {
-      this.delegate = delegate;
-      this.isOverCapacity = isOverCapacity;
-      this.limiterListener = limiterListener;
-    }
-
-    @Override public void onSuccess(Void value) {
-      try {
-        limiterListener.onSuccess(); // NOTE: limiter could block and delay the caller's callback
-        delegate.onSuccess(value);
-      } finally {
-        latch.countDown();
-      }
-    }
-
-    @Override public void onError(Throwable t) {
-      try {
-        if (isOverCapacity.test(t)) {
-          limiterListener.onDropped();
-        } else {
-          limiterListener.onIgnore();
-        }
-
-        // NOTE: the above limiter could block and delay the caller's callback
-        delegate.onError(t);
-      } finally {
-        latch.countDown();
-      }
-    }
-
-    @Override public String toString() {
-      return "LimiterReleasingCallback(" + delegate + ")";
     }
   }
 }