@Logic-32 in one review noticed we could re-use a countdown latch when
doing execute. IIRC, he mentioned `Call` is one-shot anyway, which is
true. This was good advice. Flattening `ThrottledCall` made measurably
less GC in micro benchmarks.

Before
```
Benchmark                                                           Mode     Cnt     Score     Error   Units
ThrottledCallBenchmarks.execute                                   sample  774208    17.379 ±   0.077   us/op
ThrottledCallBenchmarks.execute:execute·p0.00                     sample             1.696             us/op
ThrottledCallBenchmarks.execute:execute·p0.50                     sample            15.376             us/op
ThrottledCallBenchmarks.execute:execute·p0.90                     sample            23.712             us/op
ThrottledCallBenchmarks.execute:execute·p0.95                     sample            29.664             us/op
ThrottledCallBenchmarks.execute:execute·p0.99                     sample            54.208             us/op
ThrottledCallBenchmarks.execute:execute·p0.999                    sample            95.616             us/op
ThrottledCallBenchmarks.execute:execute·p0.9999                   sample           185.829             us/op
ThrottledCallBenchmarks.execute:execute·p1.00                     sample          5464.064             us/op
ThrottledCallBenchmarks.execute:·gc.alloc.rate                    sample      15    40.681 ±   5.778  MB/sec
ThrottledCallBenchmarks.execute:·gc.alloc.rate.norm               sample      15   552.430 ±  77.257    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space           sample      15    44.803 ±  53.001  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space.norm      sample      15   612.823 ± 727.247    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen              sample      15     0.052 ±   0.110  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen.norm         sample      15     0.704 ±   1.488    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Survivor_Space       sample      15     0.044 ±   0.184  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Survivor_Space.norm  sample      15     0.588 ±   2.434    B/op
ThrottledCallBenchmarks.execute:·gc.count                         sample      15     7.000            counts
ThrottledCallBenchmarks.execute:·gc.time                          sample      15    29.000                ms
```

After
```
Benchmark                                                     Mode       Cnti       Score   Error   Units
ThrottledCallBenchmarks.execute                               sample  606001     17.678 ±   0.429   us/op
ThrottledCallBenchmarks.execute:execute·p0.00                 sample              9.152             us/op
ThrottledCallBenchmarks.execute:execute·p0.50                 sample             14.368             us/op
ThrottledCallBenchmarks.execute:execute·p0.90                 sample             23.360             us/op
ThrottledCallBenchmarks.execute:execute·p0.95                 sample             27.904             us/op
ThrottledCallBenchmarks.execute:execute·p0.99                 sample             42.368             us/op
ThrottledCallBenchmarks.execute:execute·p0.999                sample             80.768             us/op
ThrottledCallBenchmarks.execute:execute·p0.9999               sample           3725.718             us/op
ThrottledCallBenchmarks.execute:execute·p1.00                 sample          30277.632             us/op
ThrottledCallBenchmarks.execute:·gc.alloc.rate                sample      15     30.837 ±   6.145  MB/sec
ThrottledCallBenchmarks.execute:·gc.alloc.rate.norm           sample      15    401.481 ±  60.868    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space       sample      15     32.435 ±  50.762  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space.norm  sample      15    442.977 ± 708.124    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen          sample      15      0.046 ±   0.130  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen.norm     sample      15      0.567 ±   1.615    B/op
ThrottledCallBenchmarks.execute:·gc.count                     sample      15      5.000            counts
ThrottledCallBenchmarks.execute:·gc.time                      sample      15     18.000                ms
```