@adriancole 2019-08-08
feel free to raise a different pr to improve but please use jmh or other
objective way to measure whatever it is you are concerned about

On Thu, Aug 8, 2019, 10:23 PM Logic-32 <notifications@github.com> wrote:

> Per some of my comments/concerns in the review, you may want to un-merge
> this for just a minute while we hash some things out :)
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/openzipkin/zipkin/pull/2740?email_source=notifications&email_token=AAAPVV6XVERLILATDJSTX2DQDQT7HA5CNFSM4IKA2BF2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD33Y5TI#issuecomment-519540429>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAAPVV5U5AYGJHM5R2UVNTDQDQT7HANCNFSM4IKA2BFQ>
> .
>